[WOR-1560] Upgrade to TCL 1.0.9 for MDC-aware Stairway
MdcHook is greatly reduced and renamed to StairwayLoggingHook: it now only logs at notable state transitions.
